+++ src/codec.ts
@@ -127,6 +127,7 @@ export interface PreservedMetadata {
 }
 
 export interface DecoderOptions {
+  readonly frame?: number
   readonly preserveIcc?: boolean
 }